requirements.txt
streamlit>=1.28.0
litellm>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
